Extract Locales concern
