+++ app/controllers/application_controller.rb
@@ -1,4 +1,5 @@
 class ApplicationController < ActionController::Base
+  include Locales
   include Pundit
   include StoredLocation
 
@@ -18,23 +19,6 @@ class ApplicationController < ActionController::Base
   end
 
   private
-
-  def set_locale(&action)
-    locale = if params[:locale].to_s.to_sym.in?(I18n.available_locales)
-      params[:locale]
-    else
-      I18n.default_locale
-    end
-    I18n.with_locale(locale, &action)
-  end
-
-  def default_url_options(options = {})
-    options.merge({locale: resolve_locale})
-  end
-
-  def resolve_locale(locale = I18n.locale)
-    locale == I18n.default_locale ? nil : locale
-  end
 
   def user_not_authorized
     flash[:alert] = I18n.t("pundit.errors.unauthorized")
